# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Planner JSON API
  description: |
    Mobile and programmatic access (`/api/v1`). Authenticate with an Auth0 **access token**
    whose **audience** matches `AUTH0_API_AUDIENCE` (`Authorization: Bearer &lt;token&gt;`).

    Web login uses **`AUTH0_AUDIENCE`** as the Auth0 **`audience`** parameter on `/authorize`.
    **This API** verifies Bearer tokens against **`AUTH0_API_AUDIENCE`** only.

    **Sync:** list endpoints support `updated_since`, `next_cursor`, and `limit` (default 100, max 500).

    **Optimistic concurrency:** send `If-Match: W/"&lt;updated_at&gt;"` or JSON `base_updated_at` on PATCH; on mismatch, **409** with `error.entity` (current row).

    **Idempotency (mutations):** optional `Idempotency-Key` replays cached successful responses; reuse with a different body returns **409** `IDEMPOTENCY_KEY_REUSE`.
  version: 1.0.0

servers:
  - url: /
    description: Server root

tags:
  - name: plans
  - name: tasks
  - name: resources
  - name: templates
  - name: access

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Auth0 access token (API audience).

  parameters:
    PlanID:
      name: planID
      in: path
      required: true
      schema: { type: integer, format: int64 }
    TaskID:
      name: taskID
      in: path
      required: true
      schema: { type: integer, format: int64 }
    ResourceID:
      name: resourceID
      in: path
      required: true
      schema: { type: integer, format: int64 }
    TemplateID:
      name: templateID
      in: path
      required: true
      schema: { type: integer, format: int64 }
    UpdatedSince:
      name: updated_since
      in: query
      schema: { type: string }
      description: ISO-8601-style timestamp; return rows with updated_at greater-or-equal (sync semantics).
    NextCursor:
      name: next_cursor
      in: query
      schema: { type: string }
      description: Opaque cursor from previous `next_cursor` in the response.
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 500, default: 100 }
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema: { type: string }
      description: Client-generated key; retries return the same 2xx body.
    IfMatch:
      name: If-Match
      in: header
      required: false
      schema: { type: string }
      description: Weak ETag form `W/"<updated_at>"` for concurrency (alternative to body `base_updated_at`).

  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          properties:
            code: { type: string }
            message: { type: string }
            entity: { description: Present on 409 CONFLICT (current resource) }
    Plan:
      type: object
      properties:
        id: { type: integer, format: int64 }
        name: { type: string }
        user: { type: string }
        updated_at: { type: string }
        deleted_at: { type: string, nullable: true }
    PlansListResponse:
      type: object
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/Plan' }
        next_cursor: { type: string, description: Empty if no more pages }
    CreatePlanBody:
      type: object
      required: [name]
      properties:
        name: { type: string }
    PatchPlanBody:
      type: object
      required: [name]
      properties:
        name: { type: string }
        base_updated_at: { type: string }
    Task:
      type: object
      properties:
        id: { type: integer, format: int64 }
        plan_id: { type: integer, format: int64 }
        date: { type: string }
        title: { type: string }
        subtitle: { type: string }
        description: { type: string }
        is_complete: { type: boolean }
        updated_at: { type: string }
        deleted_at: { type: string, nullable: true }
    TasksListResponse:
      type: object
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/Task' }
        next_cursor: { type: string }
    TasksByDateResponse:
      type: object
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/Task' }
    TasksWeekResponse:
      type: object
      properties:
        week: { type: string, example: "2026-20" }
        items:
          type: array
          items: { $ref: '#/components/schemas/Task' }
    CreateTaskBody:
      type: object
      required: [title, date]
      properties:
        title: { type: string }
        date: { type: string }
        subtitle: { type: string }
        description: { type: string }
        template_id: { type: integer, format: int64, description: If set, creates from template (plan ID from path ignored for source) }
    PatchTaskBody:
      type: object
      required: [title, date]
      properties:
        title: { type: string }
        date: { type: string }
        subtitle: { type: string }
        description: { type: string }
        base_updated_at: { type: string }
    PatchTaskCompleteBody:
      type: object
      properties:
        is_complete: { type: boolean }
        base_updated_at: { type: string }
    Resource:
      type: object
      properties:
        id: { type: integer, format: int64 }
        plan_id: { type: integer, format: int64 }
        title: { type: string }
        resource_type: { type: integer, format: int64 }
        content: { description: Arbitrary JSON value }
        updated_at: { type: string }
        deleted_at: { type: string, nullable: true }
    ResourcesListResponse:
      type: object
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/Resource' }
        next_cursor: { type: string }
    CreateResourceBody:
      type: object
      required: [title]
      properties:
        title: { type: string }
        resource_type: { type: integer, format: int64 }
        content: { }
    PatchResourceBody:
      type: object
      required: [title]
      properties:
        title: { type: string }
        resource_type: { type: integer, format: int64 }
        content: { }
        base_updated_at: { type: string }
    Template:
      type: object
      properties:
        id: { type: integer, format: int64 }
        plan_id: { type: integer, format: int64 }
        title: { type: string }
        subtitle: { type: string }
        description: { type: string }
        updated_at: { type: string }
        deleted_at: { type: string, nullable: true }
    TemplatesListResponse:
      type: object
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/Template' }
        next_cursor: { type: string }
    CreateTemplateBody:
      type: object
      required: [title]
      properties:
        title: { type: string }
        subtitle: { type: string }
        description: { type: string }
    PatchTemplateBody:
      type: object
      required: [title]
      properties:
        title: { type: string }
        subtitle: { type: string }
        description: { type: string }
        base_updated_at: { type: string }
    PlanAccess:
      type: object
      properties:
        id: { type: integer, format: int64 }
        plan_id: { type: integer, format: int64 }
        user: { type: string }
        updated_at: { type: string }
        deleted_at: { type: string, nullable: true }
    PlanAccessListResponse:
      type: object
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/PlanAccess' }
        next_cursor: { type: string }

security:
  - bearerAuth: []

paths:
  /api/v1/plans:
    get:
      tags: [plans]
      summary: List plans
      parameters:
        - $ref: '#/components/parameters/UpdatedSince'
        - $ref: '#/components/parameters/NextCursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PlansListResponse' }
        "401": { description: Unauthorized, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    post:
      tags: [plans]
      summary: Create plan
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreatePlanBody' }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Plan' }
        "400": { description: Bad request }
        "401": { description: Unauthorized }

  /api/v1/plans/{planID}:
    get:
      tags: [plans]
      summary: Get plan
      parameters:
        - $ref: '#/components/parameters/PlanID'
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Plan' }
        "404": { description: Not found }
    patch:
      tags: [plans]
      summary: Update plan
      parameters:
        - $ref: '#/components/parameters/PlanID'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PatchPlanBody' }
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Plan' }
        "409": { description: Conflict }
    delete:
      tags: [plans]
      summary: Soft-delete plan
      parameters:
        - $ref: '#/components/parameters/PlanID'
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        "204": { description: No content }
        "401": { description: Unauthorized }

  /api/v1/plans/{planID}/tasks:
    get:
      tags: [tasks]
      summary: List tasks by plan
      description: |
        Without `date`: incremental sync with `updated_since` / `next_cursor` / `limit`.
        With `date`: all tasks for that calendar date (no pagination cursor).
      parameters:
        - $ref: '#/components/parameters/PlanID'
        - name: date
          in: query
          schema: { type: string }
        - $ref: '#/components/parameters/UpdatedSince'
        - $ref: '#/components/parameters/NextCursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        "200":
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/TasksListResponse'
                  - $ref: '#/components/schemas/TasksByDateResponse'
    post:
      tags: [tasks]
      summary: Create task
      parameters:
        - $ref: '#/components/parameters/PlanID'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateTaskBody' }
      responses:
        "201":
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Task' }

  /api/v1/plans/{planID}/tasks/week:
    get:
      tags: [tasks]
      summary: List tasks for ISO week
      parameters:
        - $ref: '#/components/parameters/PlanID'
        - name: week
          in: query
          required: true
          schema: { type: string, example: "2026-20" }
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TasksWeekResponse' }

  /api/v1/plans/{planID}/access:
    get:
      tags: [access]
      summary: List plan sharing rows
      parameters:
        - $ref: '#/components/parameters/PlanID'
        - $ref: '#/components/parameters/UpdatedSince'
        - $ref: '#/components/parameters/NextCursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PlanAccessListResponse' }

  /api/v1/tasks/{taskID}:
    get:
      tags: [tasks]
      summary: Get task
      parameters:
        - $ref: '#/components/parameters/TaskID'
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Task' }
        "404": { description: Not found }
    patch:
      tags: [tasks]
      summary: Update task
      parameters:
        - $ref: '#/components/parameters/TaskID'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PatchTaskBody' }
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Task' }
        "409": { description: Conflict }
    delete:
      tags: [tasks]
      summary: Soft-delete task
      parameters:
        - $ref: '#/components/parameters/TaskID'
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        "204": { description: No content }

  /api/v1/tasks/{taskID}/complete:
    patch:
      tags: [tasks]
      summary: Set task completion
      parameters:
        - $ref: '#/components/parameters/TaskID'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PatchTaskCompleteBody' }
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Task' }
        "409": { description: Conflict }

  /api/v1/plans/{planID}/resources:
    get:
      tags: [resources]
      summary: List resources
      parameters:
        - $ref: '#/components/parameters/PlanID'
        - $ref: '#/components/parameters/UpdatedSince'
        - $ref: '#/components/parameters/NextCursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ResourcesListResponse' }
    post:
      tags: [resources]
      summary: Create resource
      parameters:
        - $ref: '#/components/parameters/PlanID'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateResourceBody' }
      responses:
        "201":
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Resource' }

  /api/v1/resources/{resourceID}:
    get:
      tags: [resources]
      summary: Get resource
      parameters:
        - $ref: '#/components/parameters/ResourceID'
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Resource' }
    patch:
      tags: [resources]
      summary: Update resource
      parameters:
        - $ref: '#/components/parameters/ResourceID'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PatchResourceBody' }
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Resource' }
        "409": { description: Conflict }
    delete:
      tags: [resources]
      summary: Soft-delete resource
      parameters:
        - $ref: '#/components/parameters/ResourceID'
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        "204": { description: No content }

  /api/v1/plans/{planID}/templates:
    get:
      tags: [templates]
      summary: List templates
      parameters:
        - $ref: '#/components/parameters/PlanID'
        - $ref: '#/components/parameters/UpdatedSince'
        - $ref: '#/components/parameters/NextCursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TemplatesListResponse' }
    post:
      tags: [templates]
      summary: Create template
      parameters:
        - $ref: '#/components/parameters/PlanID'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateTemplateBody' }
      responses:
        "201":
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Template' }

  /api/v1/templates/{templateID}:
    get:
      tags: [templates]
      summary: Get template
      parameters:
        - $ref: '#/components/parameters/TemplateID'
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Template' }
    patch:
      tags: [templates]
      summary: Update template
      parameters:
        - $ref: '#/components/parameters/TemplateID'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PatchTemplateBody' }
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Template' }
        "409": { description: Conflict }
    delete:
      tags: [templates]
      summary: Soft-delete template
      parameters:
        - $ref: '#/components/parameters/TemplateID'
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        "204": { description: No content }